Revert "update"
This reverts commit ecf91ab4a164deee4cb2ab5ca5f8dee9a0c824a6.

+++ app/workers/web/push_notification_worker.rb
@@ -11,13 +11,7 @@ class Web::PushNotificationWorker
 
     subscription.push(notification) unless notification.activity.nil?
   rescue Webpush::ResponseError => e
-    code = e.response.code.to_i
-
-    if (400..499).cover?(code) && ![408, 429].include?(code)
-      subscription.destroy!
-    else
-      raise e
-    end
+    subscription.destroy! if (400..499).cover?(e.response.code.to_i)
   rescue ActiveRecord::RecordNotFound
     true
   end
